feature gate the generate readme feature
