+++ Cargo.toml
@@ -16,6 +16,10 @@ exclude = [".gitignore", ".github/**"]
 name = "cargo-unleash"
 path = "src/main.rs"
 
+[features]
+default = []
+gen-readme = ["cargo-readme"]
+
 [dependencies]
 structopt = "0.3.9"
 toml_edit = "0.1.3"
@@ -27,10 +31,12 @@ reqwest = { version = "0.10.3", features = ["json"] }
 tokio = { version = "0.2.11", default_features = false }
 futures = "0.3.4"
 serde = "1.0"
-cargo-readme = "3.2"
 
 # CARGO VERSION BOUND dependencies
 cargo = "0.43.1"
 semver = "0.9.0"
 flate2 = "1.0.3"
 tar = "0.4.18"
+
+# Optional dependencies
+cargo-readme = { version="3.2", optional = true }
