+++ src/commands/mod.rs
@@ -17,3 +17,8 @@ pub use rename::rename;
 pub use set_field::set_field;
 pub use to_release::packages_to_release;
 pub use version::set_version;
+
+#[cfg(feature = "gen-readme")]
+mod gen_readme;
+#[cfg(feature = "gen-readme")]
+pub use gen_readme::gen_readme;
